+++ NEWS.md
@@ -8,6 +8,7 @@
     * `get_name()` retrieves the arrival name.
     * `get_attribute()` retrieves an attribute by name. The old method of retrieving them by providing a function with one argument is deprecated in favour of `get_attribute()`, and will be removed in version 3.7.x.
     * `get_prioritization()` retrieves the three prioritization values (`priority`, `preemptible`, `restart`) of the active arrival.
+* New shortcuts for global attributes (#110): `set_global()` and `get_global()`, equivalent to `set_attribute(global=TRUE)` and `get_attribute(global=TRUE)` respectively.
 
 ## Minor changes and fixes: